requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
diffusers>=0.25.0
transformers>=4.36.0,<4.40.0
accelerate>=0.25.0
torch>=2.0.0
Pillow>=10.0.0
numpy>=1.24.0,<2.0.0
opencv-python>=4.8.0
deep-translator>=1.11.0
